requirements.txt
langchain_google_genai
langchain
python-dotenv
together
requests
Pillow
